Use CommandPrivilege (Now IntegrationPrivilege) and remove CommandData#setDefaultEnabled
