+++ src/main/java/net/dv8tion/jda/api/interactions/commands/Command.java
@@ -21,7 +21,7 @@ import net.dv8tion.jda.api.entities.ChannelType;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.ISnowflake;
 import net.dv8tion.jda.api.interactions.commands.build.CommandData;
-import net.dv8tion.jda.api.interactions.commands.privileges.CommandPrivilege;
+import net.dv8tion.jda.api.interactions.commands.privileges.IntegrationPrivilege;
 import net.dv8tion.jda.api.requests.RestAction;
 import net.dv8tion.jda.api.requests.restaction.CommandEditAction;
 import net.dv8tion.jda.api.utils.TimeUtil;
@@ -74,10 +74,10 @@ public interface Command extends ISnowflake
     CommandEditAction editCommand();
 
     /**
-     * Retrieves the {@link CommandPrivilege CommandPrivileges} for this command.
+     * Retrieves the {@link IntegrationPrivilege IntegrationPrivileges} for this command.
      * <br>This is a shortcut for {@link Guild#retrieveCommandPrivilegesById(String)}.
      *
-     * <p>These privileges are used to restrict who can use commands through Role/User whitelists/blacklists.
+     * <p>Moderators of a guild can modify these privileges through the Integrations Menu
      *
      * <p>If there is no command with the provided ID,
      * this RestAction fails with {@link net.dv8tion.jda.api.requests.ErrorResponse#UNKNOWN_COMMAND ErrorResponse.UNKNOWN_COMMAND}
@@ -88,65 +88,11 @@ public interface Command extends ISnowflake
      * @throws IllegalArgumentException
      *         If the guild is null
      *
-     * @return {@link RestAction} - Type: {@link List} of {@link CommandPrivilege}
+     * @return {@link RestAction} - Type: {@link List} of {@link IntegrationPrivilege}
      */
     @Nonnull
     @CheckReturnValue
-    RestAction<List<CommandPrivilege>> retrievePrivileges(@Nonnull Guild guild);
-
-    /**
-     * Updates the list of {@link CommandPrivilege CommandPrivileges} for this command.
-     * <br>Note that commands are enabled by default for all members of a guild, which means you can only <em>blacklist</em> roles and members using this method.
-     * To change this behavior, use {@link CommandData#setDefaultEnabled(boolean)} on your command.
-     *
-     * <p>These privileges are used to restrict who can use commands through Role/User whitelists/blacklists.
-     *
-     * <p>If there is no command with the provided ID,
-     * this RestAction fails with {@link net.dv8tion.jda.api.requests.ErrorResponse#UNKNOWN_COMMAND ErrorResponse.UNKNOWN_COMMAND}
-     *
-     * @param  guild
-     *         The target guild from which to update the privileges
-     * @param  privileges
-     *         Complete list of {@link CommandPrivilege CommandPrivileges} for this command
-     *
-     * @throws IllegalArgumentException
-     *         If null is provided
-     * @throws IllegalStateException
-     *         If this command is not owned by this bot
-     *
-     * @return {@link RestAction} - Type: {@link List} or {@link CommandPrivilege}
-     *         The updated list of privileges for this command.
-     */
-    @Nonnull
-    @CheckReturnValue
-    RestAction<List<CommandPrivilege>> updatePrivileges(@Nonnull Guild guild, @Nonnull Collection<? extends CommandPrivilege> privileges);
-
-    /**
-     * Updates the list of {@link CommandPrivilege CommandPrivileges} for this command.
-     * <br>Note that commands are enabled by default for all members of a guild, which means you can only <em>blacklist</em> roles and members using this method.
-     * To change this behavior, use {@link CommandData#setDefaultEnabled(boolean)} on your command.
-     *
-     * <p>These privileges are used to restrict who can use commands through Role/User whitelists/blacklists.
-     *
-     * <p>If there is no command with the provided ID,
-     * this RestAction fails with {@link net.dv8tion.jda.api.requests.ErrorResponse#UNKNOWN_COMMAND ErrorResponse.UNKNOWN_COMMAND}
-     *
-     * @param  guild
-     *         The target guild from which to update the privileges
-     * @param  privileges
-     *         Complete list of {@link CommandPrivilege CommandPrivileges} for this command
-     *
-     * @throws IllegalArgumentException
-     *         If null is provided
-     * @throws IllegalStateException
-     *         If this command is not owned by this bot
-     *
-     * @return {@link RestAction} - Type: {@link List} or {@link CommandPrivilege}
-     *         The updated list of privileges for this command.
-     */
-    @Nonnull
-    @CheckReturnValue
-    RestAction<List<CommandPrivilege>> updatePrivileges(@Nonnull Guild guild, @Nonnull CommandPrivilege... privileges);
+    RestAction<List<IntegrationPrivilege>> retrievePrivileges(@Nonnull Guild guild);
 
     /**
      * Returns the {@link JDA JDA} instance of this Command
@@ -179,13 +125,6 @@ public interface Command extends ISnowflake
      */
     @Nonnull
     String getDescription();
-
-    /**
-     * Whether this command is enabled for everyone by default.
-     *
-     * @return True, if everyone can use this command by default.
-     */
-    boolean isDefaultEnabled();
 
     /**
      * The {@link Option Options} of this command.
