+++ src/main/java/net/dv8tion/jda/api/requests/restaction/CommandCreateAction.java
@@ -61,14 +61,6 @@ public interface CommandCreateAction extends RestAction<Command>, SlashCommandDa
 
     @Nonnull
     @Override
-    @Deprecated
-    @ForRemoval
-    @CheckReturnValue
-    @DeprecatedSince("5.0.0")
-    CommandCreateAction setDefaultEnabled(boolean enabled);
-
-    @Nonnull
-    @Override
     @CheckReturnValue
     CommandCreateAction setName(@Nonnull String name);
