+++ src/main/java/net/dv8tion/jda/internal/entities/GuildImpl.java
@@ -33,7 +33,7 @@ import net.dv8tion.jda.api.exceptions.ParsingException;
 import net.dv8tion.jda.api.exceptions.PermissionException;
 import net.dv8tion.jda.api.interactions.commands.Command;
 import net.dv8tion.jda.api.interactions.commands.build.CommandData;
-import net.dv8tion.jda.api.interactions.commands.privileges.CommandPrivilege;
+import net.dv8tion.jda.api.interactions.commands.privileges.IntegrationPrivilege;
 import net.dv8tion.jda.api.managers.AudioManager;
 import net.dv8tion.jda.api.managers.GuildManager;
 import net.dv8tion.jda.api.managers.GuildStickerManager;
@@ -271,7 +271,7 @@ public class GuildImpl implements Guild
 
     @Nonnull
     @Override
-    public RestAction<List<CommandPrivilege>> retrieveCommandPrivilegesById(@Nonnull String commandId)
+    public RestAction<List<IntegrationPrivilege>> retrieveCommandPrivilegesById(@Nonnull String commandId)
     {
         Checks.isSnowflake(commandId, "ID");
         Route.CompiledRoute route = Route.Interactions.GET_COMMAND_PERMISSIONS.compile(getJDA().getSelfUser().getApplicationId(), getId(), commandId);
@@ -280,65 +280,21 @@ public class GuildImpl implements Guild
 
     @Nonnull
     @Override
-    public RestAction<Map<String, List<CommandPrivilege>>> retrieveCommandPrivileges()
+    public RestAction<Map<String, List<IntegrationPrivilege>>> retrieveCommandPrivileges()
     {
         Route.CompiledRoute route = Route.Interactions.GET_ALL_COMMAND_PERMISSIONS.compile(getJDA().getSelfUser().getApplicationId(), getId());
         return new RestActionImpl<>(getJDA(), route, (response, request) -> {
-            Map<String, List<CommandPrivilege>> privileges = new HashMap<>();
+            Map<String, List<IntegrationPrivilege>> privileges = new HashMap<>();
             response.getArray().stream(DataArray::getObject).forEach(obj -> {
                 String id = obj.getString("id");
-                List<CommandPrivilege> list = parsePrivilegesList(obj);
+                List<IntegrationPrivilege> list = parsePrivilegesList(obj);
                 privileges.put(id, list);
             });
             return privileges;
         });
     }
 
-    @Nonnull
-    @Override
-    public RestAction<List<CommandPrivilege>> updateCommandPrivilegesById(@Nonnull String id, @Nonnull Collection<? extends CommandPrivilege> privileges)
-    {
-        Checks.isSnowflake(id, "ID");
-        Checks.noneNull(privileges, "Privileges");
-        Checks.check(privileges.size() <= 10, "Cannot have more than 10 privileges for a command!");
-        Route.CompiledRoute route = Route.Interactions.EDIT_COMMAND_PERMISSIONS.compile(getJDA().getSelfUser().getApplicationId(), getId(), id);
-        DataArray array = DataArray.fromCollection(privileges);
-        return new RestActionImpl<>(getJDA(), route, DataObject.empty().put("permissions", array),
-            (response, request) -> parsePrivilegesList(response.getObject()));
-    }
-
-    @Nonnull
-    @Override
-    @SuppressWarnings("deprecation")
-    public RestAction<Map<String, List<CommandPrivilege>>> updateCommandPrivileges(@Nonnull Map<String, ? extends Collection<CommandPrivilege>> privileges)
-    {
-        Checks.notNull(privileges, "Privileges");
-        privileges.forEach((key, value) -> {
-            Checks.isSnowflake(key, "Map Key");
-            Checks.noneNull(value, "Privilege List for Command");
-            Checks.check(value.size() <= 10, "Cannot have more than 10 privileges for a command!");
-        });
-        DataArray array = DataArray.empty();
-        privileges.forEach((commandId, list) -> {
-            DataObject entry = DataObject.empty();
-            entry.put("id", commandId);
-            entry.put("permissions", DataArray.fromCollection(list));
-            array.add(entry);
-        });
-
-        Route.CompiledRoute route = Route.Interactions.EDIT_ALL_COMMAND_PERMISSIONS.compile(getJDA().getSelfUser().getApplicationId(), getId());
-        return new RestActionImpl<>(getJDA(), route, RequestBody.create(Requester.MEDIA_TYPE_JSON, array.toJson()), (response, request) -> {
-            Map<String, List<CommandPrivilege>> map = new HashMap<>();
-            response.getArray().stream(DataArray::getObject).forEach(obj -> {
-                String id = obj.getString("id");
-                List<CommandPrivilege> list = parsePrivilegesList(obj);
-                map.put(id, list);
-            });
-            return map;
-        });
-    }
-
-    private List<CommandPrivilege> parsePrivilegesList(DataObject obj)
+    private List<IntegrationPrivilege> parsePrivilegesList(DataObject obj)
     {
         return obj.getArray("permissions")
                 .stream(DataArray::getObject)
@@ -346,11 +302,11 @@ public class GuildImpl implements Guild
                 .collect(Collectors.toList());
     }
 
-    private CommandPrivilege parsePrivilege(DataObject data)
+    private IntegrationPrivilege parsePrivilege(DataObject data)
     {
-        CommandPrivilege.Type type = CommandPrivilege.Type.fromKey(data.getInt("type", 1));
+        IntegrationPrivilege.Type type = IntegrationPrivilege.Type.fromKey(data.getInt("type", 1));
         boolean enabled = data.getBoolean("permission");
-        return new CommandPrivilege(type, enabled, data.getUnsignedLong("id"));
+        return new IntegrationPrivilege(this, type, enabled, data.getUnsignedLong("id"));
     }
 
     @Nonnull
