+++ src/main/java/net/dv8tion/jda/internal/requests/restaction/CommandEditActionImpl.java
@@ -89,17 +89,6 @@ public class CommandEditActionImpl extends RestActionImpl<Command> implements Co
 
     @Nonnull
     @Override
-    @Deprecated
-    @ForRemoval
-    @DeprecatedSince("5.0.0")
-    public CommandEditAction setDefaultEnabled(boolean enabled)
-    {
-        data.setDefaultEnabled(enabled);
-        return this;
-    }
-
-    @Nonnull
-    @Override
     public CommandEditAction addCheck(@Nonnull BooleanSupplier checks)
     {
         return (CommandEditAction) super.addCheck(checks);
